fix(backend): consolidate review router into statements router
- Merge review.py routes into statements.py
- Clean up main.py and routers package exports
- Update frontend API paths to match consolidated router
- Fix missing imports and lint errors in main.py
- This eliminates router shadowing and complex circular dependencies
